Refactor log API and related samples
