+++ modules/ballerina-builtin/src/main/ballerina/ballerina/log/natives.bal
@@ -1,4 +1,4 @@
-package ballerina.utils.logger;
+package ballerina.log;
 
 import ballerina.doc;
