+++ modules/ballerina-test/src/test/resources/logging.properties
@@ -22,6 +22,10 @@ java.util.logging.ConsoleHandler.level=ALL
 java.util.logging.ConsoleHandler.formatter=org.ballerinalang.logging.formatters.BallerinaLogFormatter
 org.ballerinalang.logging.formatters.BallerinaLogFormatter.format=%1$tY-%1$tm-%1$td %1$tH:%1$tM:%1$tS,%1$tL %2$s [%3$s] - %4$s %n
 
+org.ballerinalang.test.utils.TestLogHandler.level=INFO
+org.ballerinalang.test.utils.TestLogHandler.formatter=org.ballerinalang.logging.formatters.DefaultLogFormatter
+org.ballerinalang.logging.formatters.DefaultLogFormatter.format=[%1$tY-%1$tm-%1$td %1$tH:%1$tM:%1$tS,%1$tL] %2$ [%3$s] - %4$s %5$s %n
+
 ####################### LOGGERS #######################
 # Ballerina user level root logger
 ballerina.handlers=java.util.logging.ConsoleHandler
@@ -29,5 +33,5 @@ ballerina.level=ALL
 ballerina.useParentHandlers=false
 
 # JUL root logger
-.handlers=java.util.logging.ConsoleHandler
-.level=OFF
+.handlers=org.ballerinalang.test.utils.TestLogHandler
+.level=INFO
