+++ modules/ballerina-test/src/test/resources/test-src/natives/utils/logger/log-api.bal
@@ -1,4 +1,4 @@
-import ballerina.utils.logger;
+import ballerina.log;
 
 function testError (string msg) {
     logger:error(msg);
